Refactor tooltips and `merge` mode.
Simplify the way charts are merged by using direct chart refs instead of
element "merging"..

Old...
```
chart1 = new Chart({el});
chart2 = new Chart({el, merge: true});
```

New...
```
chart1 = new Chart({el});
chart2 = new Chart({parent: chart1});
```

Internally seperated the chart.tooltip options from a root chart
tooltip-view.

The tooltip-view needs to be a singleton that resides on the root chart.
The tooltip config is a per chart configuration that is used to
begin/end the lifecycle of a tooltip-view as needed.  I.e. create the
view when the first tooltip.id is detected and remove the view when the
last use of the tooltip.id is removed.

When rendering tooltips, the views are used and they iterate through all
the associated tooltip configs for each chart that have a matching id.